Backend Finishing Touches Commit: ====================================
-Added a last delete method that will be used when a user (not the owner of a photo) does not want a photo anymore and then decides to delete it. Previously th photo was deleted from the Photo entity, but now only the record in the Shared entity is deleted.

Still need to do:
-----------------

12.2. Unit testing
13. DFDs for documentation.
14. Finish all documentation.
15. Build user manual.
16. Create videos

+++ PhotoMemoriesWebSpringBoot/src/main/java/com/photomemories/web/sb/controller/PhotoController.java
@@ -96,13 +96,22 @@ public class PhotoController {
         SharedDto sharedDto = sharedCRUDService.findBySharedWithAndPhotoId(userDto.getUserId(), id);
         LOGGER.info("[Photo Controller log] deletePhoto method, access granted? {}", sharedDto.getSharedHasAccess());
 
-        if ((sharedCRUDService.checkBySharedWithAndPhotoId(email, id)) && (sharedDto.getSharedHasAccess())) {
-            photoResponse = photoCRUDService.deletePhotoDto(id, photoLink, email);
-            response = new PhotoMemoriesResponse<>(true, photoResponse);
-            return new ResponseEntity<>(response, HttpStatus.OK);
+        if (photoCRUDService.photoExists(id, photoLink)) {
+            if ((sharedCRUDService.checkBySharedWithAndPhotoId(email, id)) && (sharedDto.getSharedHasAccess())) {
+                if (sharedDto.getUserId().equals(sharedDto.getSharedWith())) {
+                    photoResponse = photoCRUDService.deletePhotoDto(id, photoLink, email);
+                } else {
+                    photoResponse = sharedCRUDService.deleteBySharedRecord(sharedDto.getSharedWith(), id, photoLink);
+                }
+                response = new PhotoMemoriesResponse<>(true, photoResponse);
+                return new ResponseEntity<>(response, HttpStatus.OK);
+            } else {
+                LOGGER.error("[Photo Controller log] deletePhoto method, access rights are not sufficient");
+                throw new SQLException("[Photo Controller log] deletePhoto method, not permitted");
+            }
         } else {
-            LOGGER.error("[Photo Controller log] deletePhoto method, access rights are not sufficient");
-            throw new SQLException("[Photo Controller log] deletePhoto method, not permitted");
+            LOGGER.error("[Photo Controller log] deletePhoto method, photo {} does not exist ", photoLink);
+            throw new SQLException("[Photo Controller Error] deletePhoto method, photo does not exist");
         }
     }
